Added Orange Map Added Awp

+++ Modern_OpenGL/Mesh.h
@@ -3,7 +3,7 @@
 #include <glm.hpp>
 #include <gl\gl.h>			// Header File For The OpenGL32 Library
 #include <gl\glu.h>			// Header File For The GLu32 Library
-
+#include "SamLoader.h"
 class Transform;
 
 struct Vertex
@@ -28,14 +28,14 @@ class Mesh
 {
 public:
 	Mesh();
-	Mesh(Vertex* vertices, unsigned short numVertices, RenderMode renderMode);
+	Mesh(Vertex* vertices, unsigned long numVertices, RenderMode renderMode);
 	virtual ~Mesh();
 
 	void Draw(Transform transform);
 
 private: 
 	Vertex* vertices;
-	unsigned short numVertices;
+	unsigned long numVertices;
 
 	RenderMode renderMode;
 };
